+++ src/main/java/um/tesoreria/guarani/hexagonal/guarani/alumno/domain/ports/in/ProcessNextPreuniversitarioUseCase.java
@@ -1,0 +1,7 @@
+package um.tesoreria.guarani.hexagonal.guarani.alumno.domain.ports.in;
+
+public interface ProcessNextPreuniversitarioUseCase {
+
+    void processNextPreuniversitario();
+
+}